requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
pymupdf
python-dotenv
openai
requests